pyproject.toml
[tool.poetry]
name = "kolzhut_rag"
version = "0.1.0"
description = "A web scraping project to collect data from kolzchut.org.il"
authors = ["Your Name <youremail@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
pymongo = "^3.12.0"
sentence-transformers = "^2.2.2"
numpy = "^2.0.0"
faiss-cpu = "^1.7.3"  # Add faiss dependency
openai = "^1.61.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"